React_EDA_Fullstack-main/Backend/requirement.txt
Django
djangorestframework
pandas
numpy
django-cors-headers
Pillow

